Audio DMA changes, tidy up of code
audio.si DMA.si

Change audio DMA to allow longer samples. Rearrange memory map to accommodate. Recompile all software using audio DMA samples. Change OUTRUN.P64 to allow selection of an audio file to play during the demo. Must be a single channel 8-bit WAV in 11025KHz.

Minor optimisations on all video output units.

cpu_functionblocks.si CPU.si ALU.si

Rearrange address generators. Reenable the 1 cycle saving on LOAD routines. Move the ALU decoder into the cpu_functionblocks.

GPU.si

Fix an error with the pixelblock, as did not account for change to the GPU clocking.

+++ RV64/SOFTWARE/c/library/PAWSlibrary.h
@@ -60,7 +60,7 @@ extern void await_beep( unsigned char );
 extern unsigned short get_beep_active( unsigned char );
 extern void tune_upload( unsigned char channel_number, unsigned short length, unsigned char *samples );
 extern void bitsample_upload_128( unsigned char channel_number, unsigned char *samples );
-extern void pcmsample_start( unsigned char channel_number, unsigned short count, const unsigned char *samples, unsigned char rate, unsigned char repeat );
+extern void pcmsample_start( unsigned char channel_number, unsigned int count, const unsigned char *samples, unsigned char rate, unsigned char repeat );
 extern void pcmsample_stop( unsigned char channel_number );
 
 // DISPLAY
